requirements.txt
Flask==3.0.1
Flask-PyMongo==2.3.0
pymongo==4.6.1
Werkzeug==3.0.1
python-dotenv==1.0.1

